// src/app/shared/models/profile.model.ts
export interface Profile {
    address: string;
    birthdate: number;
    email: string;
    email2: string;
    first_name: string;
    last_name: string;
    localid: string;
    loyalty_member_id: string;
    modified: string;
    phone: string;
    suffix: string;
    prefix: string;
    gensuffixer: string;
}